requirements.txt
langfuse==4.0.1
openai==2.29.0
pandas==3.0.1
python-dotenv==1.2.2

